requirements.txt
asgiref                   3.7.2
Django                    4.2.6
djangorestframework       3.14.0
flake8                    7.0.0
mccabe                    0.7.0
pillow                    10.2.0
pycodestyle               2.11.1
pyflakes                  3.2.0
sqlparse                  0.4.4
tzdata==2023.3
